+++ TestsFinal/Backend/Model/Operator/Multiplication.cs
@@ -14,7 +14,7 @@ namespace Backend.Model.Operator
             OperatorType = OperatorType.Multiplication;
         }
 
-        public override decimal Calculate(decimal operand1, decimal operand2)
+        public override decimal Calculate(int operand1, int operand2)
         {
             return operand1 * operand2;
         }
